refactor(groups): consolidate group-members and group-announcements into groups module
Merged group-members/ and group-announcements/ into a single groups/ module to
eliminate the cross-module route ordering constraint that required GroupMembersModule
to be registered before GroupsModule in app.module.ts. Controller ordering is now
enforced within groups.module.ts: GroupInvitationsController (literal GET /v1/groups/invitations)
is registered before GroupsController (parameterized GET /v1/groups/:id).

Also adjusts @ApiTags on all users sub-controllers to use {name, parent} hierarchy format.
