+++ apps/api/src/database/schema/index.ts
@@ -10,7 +10,7 @@
  */
 
 export * from '@/modules/assets/schema/assets.schema';
-export * from '@/modules/group-announcements/schema/group-announcements.schema';
+export * from '@/modules/groups/schema/group-announcements.schema';
 export * from '@/modules/groups/schema/group-member-stats.schema';
 export * from '@/modules/groups/schema/group-members.schema';
 export * from '@/modules/groups/schema/groups.schema';
