+++ apps/api/src/modules/groups/dto/bulk-invitation-response.dto.ts
