+++ apps/api/src/modules/groups/dto/member-response.dto.ts
