+++ apps/api/src/modules/groups/dto/my-invitation-response.dto.ts
