+++ apps/api/src/modules/groups/dto/my-membership-response.dto.ts
