+++ apps/api/src/modules/users/users-profile.controller.ts
@@ -15,7 +15,7 @@ import { UsersProfileService } from './users-profile.service';
 import { UpdateUserProfileDto } from './dto/update-user-profile.dto';
 import { UserProfileResponseDto } from './dto/user-profile-response.dto';
 
-@ApiTags('users')
+@ApiTags({ name: 'user-profile', parent: 'users' })
 @ApiBearerAuth()
 @Controller('v1/users')
 export class UsersProfileController {
